- Goes recursive on directories. - Selected file types are moved to a _meta folder in the root, instead of to the root. - File association changed from Directory to AllFileSystemObjects since it handles both files and directories. The files in the root will be moved to the _meta folder or left in place. - Configurable from registry. - Changed all SL strings to ATL strings. - Added OutputDebugStringEx function, debug output can be viewed in for example DebugView. - Added batch build files. - Optional explorer.exe restart added to PreLinkEvent in case the $(TargetPath) file write fails.
Technical discussion about folder renaming: The need for folder renaming because of a name conflict is now rare since no folders, only files, are moved to the root. It's rare that a file has the same name as a directory. The renaming function is still in place though, but renaming occurs only when needed to avoid confusing the user with foreign GUID named folders unnecessarily (in the dialog prompts and in the root after canceled move operations). Since the name conflicts are rare I suggest the renaming code is removed. The benefits are; no additional delay from the conflict check (PathFindFile), less complex code, no foreign GUID named folder whose original name may not be obvious to the user is left in the root after a canceled move operation. These benefits outweigh the downside of manual renaming on rare occasions.

Technical notes about user prompting: The "Don't move" user selection (in the move operation dialog "There is already a file with the same name in this location.") is treated more carefully by prompting the user if he still want to delete the non-empty directory to the Recycle Bin. Selecting Cancel still cancel both the move and the deletion.

Technical discussion about PreLinkEvent: To avoid restarting explorer.exe another file explorer that's less intrusive to restart can be used for testing instead. And I have plans for a Process Hacker command line option to unload the module instead, when that's available this PreLinkEvent can be removed altogether.

I'm fully committed to fix any bugs in my code. Please inform me about the issue and I will try to correct it. Please run DebugView when testing the code and provide the debug log with the issue report.