Adding recursion because that allow flattening multiple folder levels
Selected file types are moved to a _meta folder in the root, instead of to the root because that allow moving f.e. metadata files to a separate folder

File association changed from Directory to AllFileSystemObjects because it handles both files and directories. The files in the root will be moved to the _meta folder or left in place

Reading optional settings from registry

Changing SL strings to ATL strings

Adding OutputDebugStringEx function because that allow viewing debug output in DebugView

Adding batch build files

Adding optional explorer.exe restart to PreLinkEvent in case the $(TargetPath) file write fails