+ Add recursion levels option. 0 is unlimited recursion. + MetaFiles "*" interpreted as no file ending. + Original folder/file name added to (GUID) renamed folder/file to retain identifiability.
Comment to point one: 1 level (and a blank MetaFiles value) should make it work as before the previous commit.
Comment to point three: If the user select Don't move to a Move File dialog the Delete Folder dialog will be presented. That's were identifiability is simplified if the original folder name is still in the name.